requirements.txt
altair==4.2.0
attrs==22.1.0
autopep8==1.7.0
beautifulsoup4==4.11.1
blinker==1.5
blis==0.7.8
cachetools==5.2.0
catalogue==2.0.8
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
confection==0.0.2
contourpy==1.0.5
cssselect==1.1.0
cycler==0.11.0
cymem==2.0.6
dateparser==1.1.1
decorator==5.1.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0-py3-none-any.whl
entrypoints==0.4
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.8.0
fonttools==4.37.4
gitdb==4.0.9
GitPython==3.1.27
GoogleNews==1.6.5
idna==3.4
importlib-metadata==5.0.0
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
kiwisolver==1.4.4
langcodes==3.3.0
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.6.0
mpire==2.6.0
murmurhash==1.0.8
networkx==2.8.7
newspaper3k==0.2.8
nltk==3.7
numpy==1.23.3
packaging==21.3
pandas==1.5.0
pathy==0.6.2
Pillow==9.2.0
preshed==3.0.7
protobuf==3.20.3
pyarrow==9.0.0
pycodestyle==2.9.1
pydantic==1.9.2
pydeck==0.8.0b3
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.3.2
requests==2.28.1
requests-file==1.5.1
rich==12.6.0
scipy==1.9.1
semver==2.13.0
sgmllib3k==1.0.0
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
soupsieve==2.3.2.post1
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.4
streamlit==1.13.0
thinc==8.1.2
tinysegmenter==0.3
tldextract==3.3.1
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
typer==0.4.2
typing_extensions==4.3.0
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
wasabi==0.10.1
watchdog==2.1.9
zipp==3.8.1
